requirements/dev_requirements.txt
-r base_requirements.txt
colorama==0.3.7
django-debug-toolbar==1.4
lxml==3.4.4
py==1.4.31
pytest==2.9.2
requests==2.10.0
sqlparse==0.1.19
